requirements.txt
Django==3.2.8
bs4==0.0.1
beautifulsoup4==4.10.0
docx==0.2.8
Pillow==8.4.0
requests==2.26.0
tabulate==0.8.9
